ingestion/requirements.txt
# PDF Processing (minimal)
pymupdf
pymupdf4llm

# Embeddings
openai
sentence-transformers

# Vector DB
weaviate-client

# Text Processing
tiktoken

# Utilities
python-dotenv
tqdm
loguru
pydantic
pydantic-settings

# Async & HTTP
httpx
aiofiles

# PostgreSQL
asyncpg